+++ lib/ReactFauxDOM.js
@@ -1,11 +1,15 @@
 var Element = require('./Element')
 var Window = require('./Window')
-var mixin = require('./mixin')
+var core = require('./mixins/core')
+var anim = require('./mixins/core')
 
 var ReactFauxDOM = {
   Element: Element,
   defaultView: Window,
-  mixin: mixin,
+  mixins: {
+    core: core,
+    anim: anim
+  },
   createElement: function (nodeName) {
     return new Element(nodeName)
   },
